Release 0.4.5 — mypy hotfix for 0.4.4
0.4.4's remote Tests workflow (mypy --strict) failed:
SHCImpulseSwitch.impulse_length still declared -> int | None after 0.4.4
changed the underlying service to return float. Fixed to float | None.
Added a local mypy gate to scripts/local-ci.sh so this class of gap is
caught before tagging in the future (that gate didn't exist for 0.4.4).

+++ boschshcpy/models_impl.py
@@ -1018,7 +1018,7 @@ class SHCMicromoduleRelay(
             )
 
     @property
-    def impulse_length(self) -> int | None:
+    def impulse_length(self) -> float | None:
         if self._impulseswitch_service is None:
             return None
         return self._impulseswitch_service.impulse_length
